requirements.txt
certifi==2019.11.28
Click==7.0
dlib==19.19.0
face-recognition==1.2.3
face-recognition-models==0.3.0
joblib==0.14.1
Keras==2.3.1
memory-profiler==0.57.0
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mtcnn==0.1.0
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
psutil==5.7.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
wincertstore==0.2
wxPython==4.0.7.post2
XlsxWriter==1.2.7
